+++ src/rules/pycodestyle/rules/literal_comparisons.rs
@@ -4,12 +4,11 @@ use rustpython_ast::Constant;
 use rustpython_parser::ast::{Cmpop, Expr, ExprKind};
 
 use crate::ast::helpers;
-use crate::ast::helpers::{create_expr, unparse_expr};
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
 use crate::fix::Fix;
 use crate::registry::Diagnostic;
-use crate::source_code::Stylist;
+use crate::rules::pycodestyle::helpers::compare;
 use crate::violations;
 
 /// E711, E712
@@ -192,15 +191,4 @@ pub fn literal_comparisons(
     }
 
     checker.diagnostics.extend(diagnostics);
-}
-
-pub fn compare(left: &Expr, ops: &[Cmpop], comparators: &[Expr], stylist: &Stylist) -> String {
-    unparse_expr(
-        &create_expr(ExprKind::Compare {
-            left: Box::new(left.clone()),
-            ops: ops.to_vec(),
-            comparators: comparators.to_vec(),
-        }),
-        stylist,
-    )
 }
